requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.38.0
cloudinary-upload==0.0.4
dj-database-url==2.1.0
Django==5.0.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-on-heroku==1.1.2
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
whitenoise==6.6.0
